+++ frontend/src/components/categories/Categories.css
