+++ frontend/src/components/categories/Categories.test.tsx
