+++ frontend/src/components/categories/Categories.tsx
