+++ frontend/src/components/quick-actions/QuickActions.css
